ai-comments/requirements.txt
httpx>=0.27.0
openai>=1.51.0
google-genai>=0.2.0
mistralai>=1.2.0
python-dotenv>=1.0.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.3






